db2：对Where.toString的expression.value进行判空（完成）
db2：改错：java.lang.IllegalArgumentException: Can not set java.lang.Long field com.handsome.auction.bean.Auction.minPrice to java.lang.Integer
对象类型是Long，在数据库中类型是int，导致rs.getResult返回int，而Bean的对象是Long，转换出错。（完成）

db2：主键的类型支持int,Integer,long,Long四种。（完成）

+++ java_lib/src/main/java/com/wang/db2/Dao.java
@@ -14,13 +14,13 @@ public interface Dao<T> {
 
     boolean delete(Where where);
 
-    boolean deleteById(int id);
+    boolean deleteById(long id);
 
     boolean deleteAll();
 
     boolean update(T entity);
 
-    T queryById(int id);
+    T queryById(long id);
 
     List<T> queryAll();
